test: move tests to integrations
